(post-change) Update rate limiter for AlphaVantage API
Previous implementation didn't work well. It limits Job execution not an API call execution.
There were 2 options how to implement RateLimiter:
1. On HTTP level (decorate AlphaVantageAdapter)
2. On a service layer (decorate StockService)
I choose 2nd for its simplicity (as it already throws an exception on the interface level and I have a handling logic implemented).
I think the 1st option is the best. Will try to implement it in the future.